pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.17.0",
    "mysql-connector-python>=9.3.0",
    "requests>=2.32.3",
]
